Initial version of a delete command.
The delete command does not actually delete data, it just deletes and
deregisters individual snapshots. A future gc command will clean up the
data no longer referenced after a series of delete invocations.

There is currently no clean way of invoking delete, as there is no
way to get a list of snapshots to know the ID to use.

Further, this command is not yet crash-safe.