Include node- and leaf-type as salt when hashing.
These values are used when interpreting the data. This change makes it
more difficult to mistake a chunk for something else. It also allows us
to continue to store the node- and leaf-type next to the hash in the
local metadata storage without having to worry about collisions and
potential multiple interpretations (there is now 1 allowed interpretation
per hash).